deleted the index page
