+++ frontend_nextjs/styles/navheader.css
@@ -21,3 +21,10 @@
 #logout-link {
     color: var(--red);
 }
+
+#popover-notifications {
+    min-width: fit-content;
+    max-width: fit-content;
+    max-height: 40%;
+    overflow: auto;
+}
